refactor(ios): extract My Day's chip strip as DayRailView
My Day had the rail first; the Events tab should get the same surface, not a
lookalike. The strip moves to Features/Shared with a @ViewBuilder slot at
each end, so My Day keeps its expand-to-season-edge chevrons ('go far') and
the Events rail can add its own one-day steps ('go one').

Behaviour-preserving: no test changed, and the My Day screenshot is
unchanged.
